requirements.txt
streamlit
yfinance
pandas
numpy
tensorflow-cpu
joblib
matplotlib
scikit-learn
shap